+++ Source/EventFlow.TestHelpers/Test.cs
@@ -68,19 +68,25 @@ namespace EventFlow.TestHelpers
         }
 
         protected IDomainEvent<TestAggregate, TestId> ToDomainEvent<TAggregateEvent>(
-            TAggregateEvent aggregateEvent)
+            TAggregateEvent aggregateEvent,
+            int aggregateSequenceNumber = 0)
             where TAggregateEvent : IAggregateEvent
         {
             var metadata = new Metadata
+                {
+                    Timestamp = A<DateTimeOffset>()
+                };
+
+            if (aggregateSequenceNumber == 0)
             {
-                Timestamp = A<DateTimeOffset>()
-            };
+                aggregateSequenceNumber = A<int>();
+            }
 
             return DomainEventFactory.Create<TestAggregate, TestId>(
                 aggregateEvent,
                 metadata,
                 A<TestId>(),
-                A<int>(),
+                aggregateSequenceNumber,
                 A<Guid>());
         }