+++ Source/EventFlow/EventStores/EventStore.cs
@@ -190,7 +190,7 @@ namespace EventFlow.EventStores
             
             var domainEvents = await LoadEventsAsync<TAggregate, TIdentity>(id, cancellationToken).ConfigureAwait(false);
             var aggregate = await AggregateFactory.CreateNewAggregateAsync<TAggregate, TIdentity>(id).ConfigureAwait(false);
-            aggregate.ApplyEvents(domainEvents.Select(e => e.GetAggregateEvent()));
+            aggregate.ApplyEvents(domainEvents);
 
             Log.Verbose(
                 "Done loading aggregate '{0}' with ID '{1}' after applying {2} events",